pyproject.toml
[project]
name = "pytasker"
version = "0.1.5"
description = "High Level Task Executor"
authors = [
    {name = "Yair Dar", email = "yair.dar@outlook.com"},
]
license = "Apache License 2.0"

dependencies = [
    "fire>=0.4.0",
    "pydantic>=1.9.0",
    "loguru>=0.5.3",
    "pyyaml>=0.6.0",
    # "fastapi>=0.70.1",
    # "jupyter>=1.0.0",
    # "pandas>=1.3.5",
]

[project.scripts]
tasker_ctl = "pytasker.tasker_ctl:main"

[tool]
[tool.pdm]


[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

###   [tool.pytest.ini_options]
###   minversion = "6.0"
###   addopts = "-ra -q"
###   testpaths = [
###       "tests",
###       "integration",
###   ]
